map_app/requirements.txt
Flask>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
ortools>=9.0
python-dotenv>=0.19.0
